Added functions to join channel and print output.
The JoinAddonChannel joins the channel specific to this addon.  It prints out
debug information relating to the channel.  The LootGainPrint method prints
the provided message to the addon channel, if its id is not nil.  If it is nil
(for example, if the channel could not be joined), the chat message is printed
to the default chat frame.