Added functions to get static and variable player info.
The GetStaticPlayerInfo function looks up information that does not change
while a user is logged in, such as his name and class.  The
GetVariablePlayerInfo function looks up all information which could change
(but is not directly related to looting), such as the player's
specializations, level, quests, and inventory.  Both of these functions store
the collected informatin in the table parameter.