Added function AssignSourcesToUnitsList.
This function iterates through a sources list and looks for a matching unit,
based on GUID.  If a match does not exist, a basic unit is created (but the
count of mouseover units is not incremented yet).  Then, if information about
the current source's loot type has not been recorded for that unit, it is
added to the unit.  Finally, the source is given a reference to the unit, to
easily find it later.