Added functions to record loot sources.
The set of functions here handle recording data from sources and from units.
The RecordNewLootSources function takes a list of loot sources as a parameter,
and then calls RecordNewLootSource for each of them.  Thus,
RecordNewLootSources is a trivial wrapper around that function.
RecordNewLootSource, in turn, checks if the referenced unit has associated
metadata and, if so, calls RecordNewUnitSource for that unit and the loot type
(as specified in the source).  This function actually adds the information to
LootGain_Data.sources, but only if that source has not been previously
recorded.