Record mouseover info in more cases.
Do not simply record information about the moused over unit the first time it
is moused over.  The unit might be created in other cases (such as when AOE
looting that unit before ever mousing over it).  If the mouseover unit does
not have an info table when mousing over it, then populate it.