Record loot data if needed on mouseover.
A source can have loot data assigned to it but not have information known
about it, such as when AOE looting a unit before ever mousing over it.  Since
data such as name and level is not known when the loot window is opened, data
is not recorded then (but it is cached).  When that unit is later moused over
and metadata is collected, then automatically record data about that unit.