Registered CHAT_MSG_OPENING and UNIT_SPELLCAST_SUCEEDED.
The UNIT_SPELLCAST_SUCEEDED event can be used to determine when the player
completes an "opening" spell, such as Herb Gathering or Mining.  This will
allow us to determine when it was done and what spell was performed.  The
CHAT_MSG_OPENING event provides a text string which includes the spell and
also the target, so it can be used to determine the "source" name (such as
Peacebloom or Copper Vein).