Determine the source loot type.
Determining a source's loot type is somewhat complicated, and really can not
be done with only the immediate sources.  Ideally, the other loot types (or
lack thereof) should be considered.  Here, certain assumptions are used to
determine if all of the sources should be a gathering type (such as Mining or
Her Gathering).